extract layer download menu into its own component
