+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuSwitch/SwitchItem/index.tsx
@@ -1,9 +1,7 @@
 import {
   Box,
-  CircularProgress,
   createStyles,
   IconButton,
-  Menu,
   MenuItem,
   Select,
   Slider,
@@ -13,17 +11,9 @@ import {
   withStyles,
 } from '@material-ui/core';
 import OpacityIcon from '@material-ui/icons/Opacity';
-import MoreVertIcon from '@material-ui/icons/MoreVert';
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import moment from 'moment';
-import { mapValues } from 'lodash';
-import {
-  AdminLevelDataLayerProps,
-  LayerKey,
-  LayerType,
-  WMSLayerProps,
-} from '../../../../../../config/types';
+import { LayerKey, LayerType } from '../../../../../../config/types';
 import {
   getDisplayBoundaryLayers,
   LayerDefinitions,
@@ -31,8 +21,6 @@ import {
 import { clearDataset } from '../../../../../../context/datasetStateSlice';
 import { removeLayer } from '../../../../../../context/mapStateSlice';
 import {
-  dateRangeSelector,
-  layerDataSelector,
   layersSelector,
   mapSelector,
 } from '../../../../../../context/mapStateSlice/selectors';
@@ -44,14 +32,8 @@ import {
 } from '../../../../../../utils/map-utils';
 import { getUrlKey, useUrlHistory } from '../../../../../../utils/url-utils';
 import { handleChangeOpacity } from '../../../../Legends/handleChangeOpacity';
-import { LayerData } from '../../../../../../context/layers/layer-data';
-import {
-  DEFAULT_DATE_FORMAT,
-  DEFAULT_DATE_FORMAT_SNAKE_CASE,
-} from '../../../../../../utils/name-utils';
-import { downloadToFile } from '../../../../utils';
-import { castObjectsArrayToCsv } from '../../../../../../utils/csv-utils';
-import { downloadGeotiff, Extent } from '../../../../Layers/raster-utils';
+import { Extent } from '../../../../Layers/raster-utils';
+import LayerDownloadOptions from './LayerDownloadOptions';
 
 function SwitchItem({ classes, layer, extent }: SwitchItemProps) {
   const {
@@ -66,11 +48,7 @@ function SwitchItem({ classes, layer, extent }: SwitchItemProps) {
   const map = useSelector(mapSelector);
   const [isOpacitySelected, setIsOpacitySelected] = useState(false);
   const [opacity, setOpacityValue] = useState<number>(initialOpacity || 0);
-  const [isGeotiffLoading, setIsGeotiffLoading] = useState(false);
-  const [
-    downloadMenuAnchorEl,
-    setDownloadMenuAnchorEl,
-  ] = useState<HTMLElement | null>(null);
+
   const dispatch = useDispatch();
   const {
     updateHistory,
@@ -167,74 +145,6 @@ function SwitchItem({ classes, layer, extent }: SwitchItemProps) {
     toggleLayerValue(selectedId as string, true);
   };
 
-  const handleDownloadMenuClose = () => {
-    setDownloadMenuAnchorEl(null);
-  };
-
-  const handleDownloadMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
-    setDownloadMenuAnchorEl(event.currentTarget);
-  };
-
-  const { startDate: selectedDate } = useSelector(dateRangeSelector);
-  const adminLevelLayerData = useSelector(
-    layerDataSelector(layer.id, selectedDate),
-  ) as LayerData<AdminLevelDataLayerProps>;
-
-  const getFilename = (): string => {
-    const safeTitle = layer.title ?? layer.id;
-    if (selectedDate && (layer as AdminLevelDataLayerProps).dates) {
-      const dateString = moment(selectedDate).format(
-        DEFAULT_DATE_FORMAT_SNAKE_CASE,
-      );
-      return `${safeTitle}_${dateString}`;
-    }
-    return safeTitle;
-  };
-
-  const handleDownloadGeoJson = (): void => {
-    downloadToFile(
-      {
-        content: JSON.stringify(adminLevelLayerData.data.features),
-        isUrl: false,
-      },
-      getFilename(),
-      'application/json',
-    );
-    handleDownloadMenuClose();
-  };
-
-  const handleDownloadCsv = (): void => {
-    const translatedColumnsNames = mapValues(
-      adminLevelLayerData.data.layerData[0],
-      (v, k) => (k === 'value' ? t(adminLevelLayerData.layer.id) : t(k)),
-    );
-    downloadToFile(
-      {
-        content: castObjectsArrayToCsv(
-          adminLevelLayerData.data.layerData,
-          translatedColumnsNames,
-          ';',
-        ),
-        isUrl: false,
-      },
-      getFilename(),
-      'text/csv',
-    );
-    handleDownloadMenuClose();
-  };
-
-  const handleDownloadGeoTiff = (): void => {
-    setIsGeotiffLoading(true);
-    downloadGeotiff(
-      (layer as WMSLayerProps).serverLayerName,
-      extent,
-      moment(selectedDate).format(DEFAULT_DATE_FORMAT),
-      dispatch,
-      () => setIsGeotiffLoading(false),
-    );
-    handleDownloadMenuClose();
-  };
-
   const menuTitle = group ? (
     <>
       <Typography className={selected ? classes.title : classes.titleUnchecked}>
@@ -292,42 +202,11 @@ function SwitchItem({ classes, layer, extent }: SwitchItemProps) {
         >
           <OpacityIcon />
         </IconButton>
-        <IconButton
-          disabled={!selected || isGeotiffLoading}
-          onClick={handleDownloadMenuOpen}
-        >
-          <MoreVertIcon />
-        </IconButton>
-        {isGeotiffLoading && (
-          <Box display="flex" alignItems="center">
-            <CircularProgress size="20px" />
-          </Box>
-        )}
-        <Menu
-          id="download-menu"
-          anchorEl={downloadMenuAnchorEl}
-          keepMounted
-          open={Boolean(downloadMenuAnchorEl)}
-          onClose={handleDownloadMenuClose}
-        >
-          <MenuItem onClick={handleDownloadMenuClose}>
-            Exposure Analysis
-          </MenuItem>
-          {layer.type === 'admin_level_data' && (
-            <>
-              <MenuItem onClick={handleDownloadCsv}>Download as CSV</MenuItem>
-              <MenuItem onClick={handleDownloadGeoJson}>
-                Download as GeoJSON
-              </MenuItem>
-            </>
-          )}
-          {layer.type === 'wms' &&
-            layer.baseUrl.includes('api.earthobservation.vam.wfp.org/ows') && (
-              <MenuItem onClick={handleDownloadGeoTiff}>
-                Download as GeoTIFF
-              </MenuItem>
-            )}
-        </Menu>
+        <LayerDownloadOptions
+          layer={layer}
+          extent={extent}
+          selected={selected}
+        />
       </Box>
       {selected && isOpacitySelected && (
         <Box display="flex" justifyContent="right" alignItems="center">
